Added sequences and display freature
And this is where it all starts to come together. This changeset
adds the sending and resending of sequences instead of just single
letters. This means that all the little sliders in the UI that I
started on all that time ago now work!

Users can also now view their answers. With the app in its current
state, users can select a set of letters, play and replay the morse
for them, then reveal the answer.